* todos los tipos pasados a stdint compliant(ie uint64_t, uint16_t, etc) * puesto const, static y volatile donde va * preparada lib de comandos para diferentes privilegios en el futuro.

+++ DeliriOS_64bits/include/multicore.h
@@ -32,35 +32,35 @@ typedef enum {
 //MultiProcessor Configuration Table Entry for Processor. Table 4.4 of spec
 typedef struct {
 	//ID for Local APIC of processor
-	uchar local_apic_id;
+	uint8_t local_apic_id;
 	//Version register number
-	uchar version;
+	uint8_t version;
 	//Enabled bit. Zero if unusable
-	uchar enabled : 1;
+	uint8_t enabled : 1;
 	//Bootstrap processor. One if it is
-	uchar bootstrap : 1;
+	uint8_t bootstrap : 1;
 	//Reserved bits
-	uchar __reserved : 6;
+	uint8_t __reserved : 6;
 	//CPU Signature
 	//	Contains in consecutive order stepping (4 bits), model (4),	family (4)
 	//	All following bits are reserved.
 	//
 	//	Table 4.5 has values.
-	uchar stepping : 4;
-	uchar model : 4;
-	uchar family : 4;
+	uint8_t stepping : 4;
+	uint8_t model : 4;
+	uint8_t family : 4;
 	//Reserved bits of CPU signature flag
-	uchar __reserved_signature_high_nibble : 4;
-	uchar __reserved_signature_byte : 4;
-	ushort __reserved_signature_word;
+	uint8_t __reserved_signature_high_nibble : 4;
+	uint8_t __reserved_signature_byte : 4;
+	uint16_t __reserved_signature_word;
 	//Feature flags as returned by CPUID instruction. Table 4.6 has values.
-	uint features;
+	uint64_t features;
 } __attribute__((__packed__)) processor_entry;
 
 //MultiProcessor Configuration Table Entry for Bus. Table 4.7 of spec.
 typedef struct {
 	//Identifier. BIOS assigns identifiers starting from zero.
-	uchar id;
+	uint8_t id;
 	//String identifying the type of bus;
 	//Table 4.8 has possible values.
 	char type[6];
@@ -68,11 +68,11 @@ typedef struct {
 
 //MultiProcessor Configuration Table Entry for IOAPIC. Table 4.9 of Spec
 typedef struct {
-	uchar id;
+	uint8_t id;
 	//Version Register of APIC
-	uchar version : 7;
+	uint8_t version : 7;
 	//Enabled bit. If zero the APIC is unusable
-	uchar enabled : 1;
+	uint8_t enabled : 1;
 	//Base address of APIC
 	void * base_address;
 } __attribute__((__packed__)) ioapic_entry;
@@ -81,44 +81,44 @@ typedef struct {
 //	Table 4.10 of Spec
 typedef struct {
 	//Type of interrupt. Values in table 4-11
-	uchar type;
+	uint8_t type;
 	//Polarity of input signals. Table 4-10 has values.
-	uchar polarity : 2;
+	uint8_t polarity : 2;
 	//Trigger mode of input signals. Table 4-10 has values.
-	uchar trigger_mode : 2;
+	uint8_t trigger_mode : 2;
 	//Identifies the bus from where this interrupt comes from.
-	uchar source_bus_id;
+	uint8_t source_bus_id;
 	//Identifies the interrupt from bus. Starts at 0
-	uchar source_bus_irq;
+	uint8_t source_bus_irq;
 	//Destination IO APIC Id
-	uchar dest_apic_id;
+	uint8_t dest_apic_id;
 	//Destination IO APIC INTIn
-	uchar dest_apic_intin;
+	uint8_t dest_apic_intin;
 } __attribute__((__packed__)) intr_assign_entry;
 
 //MultiProcessor Configuration Table Entry for Local IO Interrupts.
 //	Table 4.12 of Spec
 typedef struct {
 	//Type of interrupt. Values in table 4-11
-	uchar type;
+	uint8_t type;
 	//Polarity of input signals. Table 4-10 has values.
-	uchar polarity : 2;
+	uint8_t polarity : 2;
 	//Trigger mode of input signals. Table 4-10 has values.
-	uchar trigger_mode : 2;
+	uint8_t trigger_mode : 2;
 	//Identifies the bus from where this interrupt comes from.
-	uchar source_bus_id;
+	uint8_t source_bus_id;
 	//Identifies the interrupt from bus. Starts at 0
-	uchar source_bus_irq;
+	uint8_t source_bus_irq;
 	//Destination IO APIC Id
-	uchar dest_apic_id;
+	uint8_t dest_apic_id;
 	//Destination IO APIC LINTIn
-	uchar dest_apic_lintin;
+	uint8_t dest_apic_lintin;
 } __attribute__((__packed__)) local_intr_assign_entry;
 
 // MultiProcessor Configuration Table Entry in General.
 typedef struct {
 	//Entry type is a byte. Enums are ints by default in C. So no enum here.
-	uchar entry_type;
+	uint8_t entry_type;
 	union {
 		processor_entry			processor;
 		ioapic_entry			ioapic;
@@ -133,12 +133,12 @@ typedef struct mp_config_table{
 	//Signature: Should be PCMP
 	char signature[4];
 	//Length of the base configuration in bytes, including header.
-	ushort length;
+	uint16_t length;
 	//Revision of spec.  1 for 1.1, 4 for 1.4
-	uchar version;
+	uint8_t version;
 	//Checksum of the base configuration table.
 	//All bytes including checksum bytes must equate zero.
-	uchar checksum;
+	uint8_t checksum;
 	//OEM ID: Name of manufacturer of system hardware. Not NULL Terminated.
 	char oem_id[8];
 	//Product ID: Name of product family of the system. Not NULL Terminated.
@@ -146,18 +146,18 @@ typedef struct mp_config_table{
 	//OEM Table Pointer: Optional OEM defined config table. Zero if not defined.
 	struct mp_config_table * oem_config_table;
 	//OEM Table Size: Length of optional OEM table. Zero if not defined.
-	ushort oem_table_length;
+	uint16_t oem_table_length;
 	//Entries following this base header in memory.
-	ushort entry_count;
+	uint16_t entry_count;
 	//Base address by which each processor acceses the local apic. 
-	uint local_apic_addr;
+	uint64_t local_apic_addr;
 	//Length in bytes of the extended table entries. Zero if there are none.
-	ushort extended_table_length;
+	uint16_t extended_table_length;
 	//Extended table checksum. All bytes of the extended table must sum to this
 	//value. Zero if there are no extended entries.
-	uchar extended_table_checksum;
+	uint8_t extended_table_checksum;
 	//Reserved space according to spec
-	uchar __reserved;
+	uint8_t __reserved;
 	//Consecutive ENTRY_COUNT entries
 	mp_entry entries[];
 } __attribute__((__packed__)) mp_config_table;
@@ -170,20 +170,20 @@ typedef struct {
 	//configuration. All zeros if the configuration table does not exist.
 	mp_config_table * config;
 	//Length: Number of 16 byte chunks of this structure. Should be 1
-	uchar length;
+	uint8_t length;
 	//Version: Number of the MP Specification used. 1 for 1.1, 4 for 1.4
-	uchar version;
+	uint8_t version;
 	//Checksum: Sum of all the bytes in the struct should be cero.
-	uchar checksum;
+	uint8_t checksum;
 	//MP Features 1: Feature flags. When zero, indicates that a configuration
 	//table is present. If not, indicates which default configuration the
 	//system implements. See table 4-1 of the spec.
-	uchar mp_features1;
+	uint8_t mp_features1;
 	//MP Features 2: Bits 0-6 are reserved. Bit 7 indicates IMCR is present
 	//and PIC Mode is implemented. Otherwise Virtual Wire Mode is implemented.
-	uchar mp_features2;
+	uint8_t mp_features2;
 	//MP Features 3: Reserved, must be zero.
-	uchar mp_features3[3];
+	uint8_t mp_features3[3];
 } __attribute__((__packed__)) mp_float_struct;
 
 //IMCRP bit: Determines if Interrupt Mode Configuration Register is present.
@@ -193,12 +193,12 @@ typedef struct {
 //Interrupt Command Register (ICR). See Section 10.6
 typedef struct {
 	//The vector number of the interrupt being sent.
-	uchar vector; 
+	uint8_t vector; 
 	//Specifies the type of IPI to be sent.
-	uchar delivery_mode:3; 
+	uint8_t delivery_mode:3; 
 	//Selects either physical (0) or logical (1) destination mode 
 	//(see Section 10.6.2, “Determining IPI Destination”).
-	uchar destination_mode:1;		
+	uint8_t destination_mode:1;		
 	// Indicates the IPI delivery status: 
 	//
 	// 0(idle) indicates the local 
@@ -206,33 +206,33 @@ typedef struct {
 	//
 	// 1 (send pending) Indicates that this local apic has not completed 
 	// sending the last IPI.
-	uchar delivery_status:1; 
-	uchar __reserved1:1;
+	uint8_t delivery_status:1; 
+	uint8_t __reserved1:1;
 	//For the INIT level de-assert delivery mode this flag must be set to 0; 
 	//for all other delivery modes it must be set to 1.
-	uchar level:1; 
+	uint8_t level:1; 
 	//Selects the trigger mode when using the INIT level de-assert delivery 
 	//mode: edge (0) or level (1)
-	uchar trigger_mode:1; 
-	uchar __reserved2:2;
+	uint8_t trigger_mode:1; 
+	uint8_t __reserved2:2;
 	// Shorthands are defined for the following cases: 
 	//	No shorthand(00) (Destination specified in destination field),
 	//	Software self interrupt (01), 
 	//	IPIs to all processors in the system including the sender (10), 
 	//	IPIs to all processors in the system excluding the sender (11).
-	uchar destination_shorthand:2; 
+	uint8_t destination_shorthand:2; 
 	//Specifies the target processor or processors.
 	//This field is only used when the destination shorthand field is set 
 	//to 00B.	
-	uchar __reserved3 :4;
-	uchar __reserved4[4];
+	uint8_t __reserved3 :4;
+	uint8_t __reserved4[4];
 	// Destination field
 	//	00: if the destination mode is physical, the destination field
 	//		contains the APIC ID of the destination
 	//
 	//	when the mode is logical, the interpretation of this field can be
 	//	found in Intel SDM Vol 3 Chapter 10
-	uchar destination_field; 
+	uint8_t destination_field; 
 } __attribute__((__packed__)) intr_command_register;
 
 typedef enum {
